+++ src/main/java/com/spring/repository/SpecialRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SpecialRepository extends JpaRepository<Special, Long> {
+public interface SpecialRepository extends JpaRepository<Special, Integer> {
